noop: move sha2 code to a separate file
